% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gnaf.R
\name{build_gnaf}
\alias{build_gnaf}
\title{Builds G-NAF into a flat \code{data.table} object.}
\usage{
build_gnaf(
  setup = gnaf_setup_data,
  simple = FALSE,
  drop_empty_vars = TRUE,
  drop_date_variables = TRUE,
  add_geo_coordinate = TRUE,
  add_meshblock_2011 = TRUE,
  add_meshblock_2016 = TRUE,
  add_locality_alias_variants = FALSE,
  add_alias_links = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{setup}{defining the default \code{gnaf_setup_data} object.}

\item{simple}{A logical argument to remove (potentially) unused, technical fields. Default is \code{FALSE}.}

\item{drop_empty_vars}{A logical argument to remove empty fields. Default is \code{TRUE}.}

\item{drop_date_variables}{A logical argument to remove date (creation / retire / modification) fields. Default is \code{TRUE}.}

\item{add_geo_coordinate}{A logical argument to add geographical coordinate latitude and longitude fields. Default is \code{TRUE}.}

\item{add_meshblock_2011}{A logical argument to Mesh Block 2011 fields. Default is \code{TRUE}.}

\item{add_meshblock_2016}{A logical argument to Mesh Block 2016 fields. Default is \code{TRUE}.}

\item{add_locality_alias_variants}{A logical argument to include locality alias variants. Default is \code{FALSE}.}

\item{add_alias_links}{A logical argument to allow inclusion of primary id on alias addresses. Default is \code{FALSE}.}

\item{verbose}{A logical argument to verbose.  Default is \code{TRUE}.}
}
\value{
Returns a flat \code{data.table} object of Australian (G-NAF) address information.
}
\description{
Builds G-NAF into a flat \code{data.table} object.
}
\examples{
# Update
}
\author{
Kyle Haynes, \email{kyle@kylehaynes.com.au}.
}
