% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_street_address.R
\name{create_street_address}
\alias{create_street_address}
\title{Creates street address strings from G-NAF.}
\usage{
create_street_address(gnaf_data, string_var = "STREET_ADDRESS", verbose = TRUE)
}
\arguments{
\item{gnaf_data}{G-NAF data.table input.}

\item{string_var}{A character string for the output street address string. Default is \code{FALSE}.}
}
\value{
Returns a new column on an existing G-NAF data.table.
}
\description{
Creates street address strings from G-NAF.
}
\examples{
# Update
}
\author{
Kyle Haynes, \email{kyle@kylehaynes.com.au}.
}
