% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_address_alias}
\alias{get_address_alias}
\title{Returns the address_alias information.}
\usage{
get_address_alias(setup = gnaf_setup_data)
}
\arguments{
\item{setup}{A character vector identifying the setup object run by calling \code{setup()}.}
}
\value{
Returns \emph{\strong{UPDATE}}
}
\description{
Returns address_alias information.
}
\examples{
# Update
}
\author{
Kyle Haynes, \email{kyle@kylehaynes.com.au}.
}
