% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_address_mesh_block}
\alias{get_address_mesh_block}
\title{Returns address_mesh_block data.}
\usage{
get_address_mesh_block(
  setup = gnaf_setup_data,
  years = c("2021", "2016", "2011")
)
}
\arguments{
\item{setup}{A character vector identifying the setup object run by calling \code{setup()}.}
}
\value{
Returns \emph{\strong{UPDATE}}
}
\description{
Returns address_mesh_block data.
}
\examples{
# Update
}
\author{
Kyle Haynes, \email{kyle@kylehaynes.com.au}.
}
