% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gnaf.R
\name{get_gnaf}
\alias{get_gnaf}
\title{Download and unpack the latest version of G-NAF.}
\usage{
get_gnaf(
 
    url = "https://data.gov.au/data/dataset/19432f89-dc3a-4ef3-b943-5326ef1dbecc/resource/4b084096-65e4-4c8e-abbe-5e54ff85f42f/download/nov20_gnaf_pipeseparatedvalue.zip",
  dest_folder = "./",
  recursive = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{url}{A character vector identifying the root directory of where the G-NAF extract is
located. Note that it will always download the most recent gnaf, despite the default URL
indicating Nov 2020}

\item{dest_folder}{Path where G-NAF will be downloaded to.  Default is \code{"./"}.}

\item{recursive}{If file paths should allow to be recursively created. Default is \code{TRUE}.}

\item{verbose}{A logical argument to determine if the function should be verbose or not. Default is \code{TRUE}.}
}
\value{
Returns The latest version of G-NAF.
}
\description{
Convenience function for downloading and unpacking G-NAF
}
\author{
Kyle Haynes, \email{kyle@kylehaynes.com.au}.
}
