% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_mb_2016}
\alias{get_mb_2016}
\title{Returns mb_2016 information.}
\usage{
get_mb_2016(setup = gnaf_setup_data)
}
\arguments{
\item{setup}{A character vector identifying the setup object run by calling \code{setup()}.}
}
\value{
Returns \emph{\strong{UPDATE}}
}
\description{
Returns mb_2016 information.
}
\examples{
# Update
}
\author{
Kyle Haynes, \email{kyle@kylehaynes.com.au}.
}
