% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_primary_secondary}
\alias{get_primary_secondary}
\title{Returns a list of primary and secondary connections.}
\usage{
get_primary_secondary(setup = gnaf_setup_data)
}
\arguments{
\item{setup}{A character vector identifying the setup object run by calling \code{setup()}.}
}
\value{
Returns \emph{\strong{UPDATE}}
}
\description{
Returns a list of primary and secondary connections.
}
\examples{
# Update
}
\author{
Kyle Haynes, \email{kyle@kylehaynes.com.au}.
}
