% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_state}
\alias{get_state}
\title{Returns the get_state information.}
\usage{
get_state(setup = gnaf_setup_data)
}
\arguments{
\item{setup}{A character vector identifying the setup object run by calling \code{setup()}.}
}
\value{
Returns \emph{\strong{UPDATE}}
}
\description{
Returns state information.
}
\examples{
# Update
}
\author{
Kyle Haynes, \email{kyle@kylehaynes.com.au}.
}
