% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_street_locality_alias}
\alias{get_street_locality_alias}
\title{Returns a list of locality alias'.}
\usage{
get_street_locality_alias(setup = gnaf_setup_data)
}
\arguments{
\item{setup}{A character vector identifying the setup object run by calling \code{setup()}.}
}
\value{
Returns \emph{\strong{UPDATE}}
}
\description{
Returns a list of locality alias'.
}
\examples{
# Update
}
\author{
Kyle Haynes, \email{kyle@kylehaynes.com.au}.
}
