% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_data.R
\name{search_data}
\alias{search_data}
\title{Shiny convenience GUI for searching data}
\usage{
search_data(d)
}
\arguments{
\item{d}{data.frame type object}
}
\description{
Shiny convenience GUI for searching data
}
\author{
Kyle Haynes, \email{kyle@kylehaynes.com.au}.
}
